+++ src/math/Tuple.cpp
@@ -12,21 +12,21 @@ import rt.utils;
 namespace rt {
 
     // --- Tuple Struct Implementations ---
-    Tuple::Tuple() : x(0.0), y(0.0), z(0.0), w(0.0) {}
-    Tuple::Tuple(double _x, double _y, double _z, double _w)
+    Tuple::Tuple() : x(0.0f), y(0.0f), z(0.0f), w(0.0f) {}
+    Tuple::Tuple(float _x, float _y, float _z, float _w)
         : x(_x), y(_y), z(_z), w(_w) {}
 
-    Tuple Tuple::operator*(double scalar) const {
+    Tuple Tuple::operator*(float scalar) const {
         return {x * scalar, y * scalar, z * scalar, w * scalar};
     }
 
-    Tuple Tuple::operator/(double scalar) const {
-        assert(scalar != 0.0 && "Attempted to divide a Tuple by zero");
-        double scalar_inv = 1.0 / scalar;
+    Tuple Tuple::operator/(float scalar) const {
+        assert(scalar != 0.0f && "Attempted to divide a Tuple by zero");
+        float scalar_inv = 1.0f / scalar;
         return {x * scalar_inv, y * scalar_inv, z * scalar_inv, w * scalar_inv};
     }
 
-    double Tuple::operator[](std::size_t index) const {
+    float Tuple::operator[](std::size_t index) const {
         switch (index) {
             case 0: return x;
             case 1: return y;
@@ -36,7 +36,7 @@ namespace rt {
         }
     }
 
-    double& Tuple::operator[](std::size_t index) {
+    float& Tuple::operator[](std::size_t index) {
         switch (index) {
             case 0: return x;
             case 1: return y;
@@ -58,16 +58,16 @@ namespace rt {
     }
 
     // --- Derived Types ---
-    Point::Point() : Tuple(0.0, 0.0, 0.0, 1.0) {}
-    Point::Point(double x, double y, double z) : Tuple(x, y, z, 1.0) {}
+    Point::Point() : Tuple(0.0f, 0.0f, 0.0f, 1.0f) {}
+    Point::Point(float x, float y, float z) : Tuple(x, y, z, 1.0f) {}
     Point::Point(const Tuple& t) : Tuple(t) {
-        assert(equal(t.w, 1.0));
+        assert(equal(t.w, 1.0f));
     }
 
-    Vector::Vector() : Tuple(0.0, 0.0, 0.0, 0.0) {}
-    Vector::Vector(double x, double y, double z) : Tuple(x, y, z, 0.0) {}
+    Vector::Vector() : Tuple(0.0f, 0.0f, 0.0f, 0.0f) {}
+    Vector::Vector(float x, float y, float z) : Tuple(x, y, z, 0.0f) {}
     Vector::Vector(const Tuple& t) : Tuple(t) {
-        assert(equal(t.w, 0.0));
+        assert(equal(t.w, 0.0f));
     }
 
     // --- Geometry Overloads ---
@@ -99,26 +99,26 @@ namespace rt {
         return {-v.x, -v.y, -v.z};
     }
 
-    Vector operator*(const Vector& v, double scalar) {
+    Vector operator*(const Vector& v, float scalar) {
         return {v.x * scalar, v.y * scalar, v.z * scalar};
     }
 
-    Vector operator*(double scalar, const Vector& v) {
+    Vector operator*(float scalar, const Vector& v) {
         return v * scalar;
     }
 
-    Vector operator/(const Vector& v, double scalar) {
-        assert(scalar != 0.0 && "Attempted to divide a Vector by zero");
-        double inv = 1.0 / scalar;
+    Vector operator/(const Vector& v, float scalar) {
+        assert(scalar != 0.0f && "Attempted to divide a Vector by zero");
+        float inv = 1.0f / scalar;
         return {v.x * inv, v.y * inv, v.z * inv};
     }
 
     // --- Factory Utilities ---
-    Point createPoint(double x, double y, double z) {
+    Point createPoint(float x, float y, float z) {
         return {x, y, z};
     }
 
-    Vector createVector(double x, double y, double z) {
+    Vector createVector(float x, float y, float z) {
         return {x, y, z};
     }
 
@@ -127,16 +127,16 @@ namespace rt {
         return -a;
     }
 
-    double getVectorMagnitude(const Vector& a) {
+    float getVectorMagnitude(const Vector& a) {
         return std::sqrt(a.x * a.x + a.y * a.y + a.z * a.z);
     }
 
     Vector normalizeVector(const Vector& a) {
-        double inv_mag = 1.0 / getVectorMagnitude(a);
+        float inv_mag = 1.0f / getVectorMagnitude(a);
         return createVector(a.x * inv_mag, a.y * inv_mag, a.z * inv_mag);
     }
 
-    double dotProduct(const Vector& a, const Vector& b) {
+    float dotProduct(const Vector& a, const Vector& b) {
         return (a.x * b.x + a.y * b.y + a.z * b.z + a.w * b.w);
     }
 
@@ -150,7 +150,7 @@ namespace rt {
     }
 
     Vector reflect(const Vector& in, const Vector& normal) {
-        return in - normal * 2.0 * dotProduct(in, normal);
+        return in - normal * 2.0f * dotProduct(in, normal);
     }
 
 } // namespace rt
